fix(trips) : Rename dto files with appropriate names
