+++ src/trip/dto/create-trips-data.ts
