+++ src/trip/dto/update-trips-data.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateTripDto } from './create-trips-data';
+
+export class UpdateTripDto extends PartialType(CreateTripDto) {}
