--- src/trip/dto/update-trips.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateTripDto } from './create-trips.dto';
-
-export class UpdateTripData extends PartialType(CreateTripDto) {}
